Move HW wallets integration tests into separate module
