+++ lib/core-integration/cardano-wallet-core-integration.cabal
@@ -70,6 +70,7 @@ library
       Test.Integration.Scenario.API.Network
       Test.Integration.Scenario.API.Transactions
       Test.Integration.Scenario.API.Wallets
+      Test.Integration.Scenario.API.HWWallets
       Test.Integration.Scenario.CLI.Addresses
       Test.Integration.Scenario.CLI.Miscellaneous
       Test.Integration.Scenario.CLI.Mnemonics
